feat(decoration): add blur effect and adjust active/inactive opacity
Adjusts the active and inactive opacity levels and adds a blur effect to the decoration module.

BREAKING CHANGE: The default height of the groupbar has been increased from 12 to 14, and the stacked property has been set to false. Additionally, the disable_hyprland_logo option has been set to true, and the font_family option has been set to Fira Code.