feat(shell): introduce plugin support for zsh
This commit introduces a plugin system for the zsh shell. This allows users to easily manage and install plugins without having to manually edit their `.zshrc` file.

The plugin system is implemented as a set of Nix expressions in the `src/apps/shell/zsh/plugins` directory. Each expression defines a plugin and its dependencies. Users can then install plugins by adding the corresponding expression to their `configuration.nix` file.

Once a plugin is installed, it will be automatically loaded when the user starts zsh. Plugins can define custom commands, key bindings, and other features to extend the functionality of zsh.

The plugin system is still in its early stages of development, but it has the potential to significantly improve the usability and extensibility of zsh.