feat(memory): introduce zram support
Adds support for ZRAM swap in the default system configuration.

Specifically, this commit introduces a new `memory` attribute to the `default` NixOS configuration, which can be used to enable and configure ZRAM swap.
The default value of this attribute, as specified in the newly created `src/systems/memory/default.nix`, is to enable ZRAM swap with a maximum memory size of 20 GB, a compression algorithm of "zstd", and a single swap device.