feat(hyprland): replace activation to download file
This commit replaces the activation script with a file source to download the wallpaper to the correct location.

The previous activation script did not work because it copied the wallpaper to the wrong location. This change fixes that issue by using a file source to download the wallpaper to the correct location.