fix(hyprland): lower timeout values in hypridle
This commit lowers the timeout values in hypridle to make the screensaver activate sooner.

The timeout values were previously set to 5 minutes and 15 minutes. This commit changes them to 15 minutes and 60 minutes, respectively.

This change was made to improve the user experience by making the screensaver activate sooner when the user is idle.