feat(apps): add yazi as file manager
Adds yazi as the default file manager.

This commit introduces yazi as the new file manager for the system, replacing the previous file manager. Yazi is a powerful and extensible file manager with a number of features that make it a great choice for our system, including:

* Advanced file search and filtering
* Customizable keyboard shortcuts
* Support for a variety of file formats
* Extensibility through plugins

This change also updates the MIME associations to use the appropriate applications for different file types.