feat: added config for xdg mime (#2)
* feat: enable p7zip in yazi

Add p7zip to yazi, and enable yazi on zsh by default.

* feat(apps): add yazi as file manager

Adds yazi as the default file manager.

This commit introduces yazi as the new file manager for the system, replacing the previous file manager. Yazi is a powerful and extensible file manager with a number of features that make it a great choice for our system, including:

* Advanced file search and filtering
* Customizable keyboard shortcuts
* Support for a variety of file formats
* Extensibility through plugins

This change also updates the MIME associations to use the appropriate applications for different file types.

* fix(window_rules): float and pin Gemini

Gemini is a small window that doesn't behave like other windows.
It floats, stays on top of all windows, and can only be resized by dragging any edge of the window.
This commit adds rules to float, pin, and size Gemini.

* fix(container/docker): update dataRootDir to /media/Containers/docker

Update dataRootDir for docker to /media/Containers/docker to match convention.

* refactor: add impala, nixbld

Adds impala networking and nixbld user to default system config.