docs: update go GOME path
Revise the GOME path to be more consistent with other language environments.

BREAKING CHANGE: The previous GOME path, which was only used for aliasing go commands, has been deprecated and replaced with a new GOME path that is more consistent with other language environments. This change should not affect any existing usage of go commands, but it may require updating any custom scripts or aliases that rely on the old GOME path.