refactor: switch pulseaudio setup in hardware config to force unsetting
This commit refactors the pulseaudio setup in the hardware config to force unsetting, allowing a custom pulseaudio setup in the home manager config to take precedence.

Fixes a race condition where the hardware config would set pulseaudio to true before the home manager config could unset it, resulting in pulseaudio being enabled even when the user desired it to be disabled.